The track list is now a completely separated view to enable the app to set it when clicking on the corresponding button. Refactoring a lot of code, and adding comments.
