+++ src/main/resources/fxml/main-view.fxml
@@ -37,27 +37,6 @@
       </VBox>
    </left>
    <center>
-      <VBox alignment="TOP_CENTER" style="-fx-border-width: 1; -fx-border-color: gray; -fx-border-style: hidden solid hidden solid;" BorderPane.alignment="CENTER">
-         <padding>
-            <Insets left="25.0" right="25.0" />
-         </padding>
-         <Label styleClass="h1" text="Vos musiques" />
-         <Separator prefWidth="200.0" styleClass="separator-dark" />
-         <TableView fx:id="musicsList" VBox.vgrow="ALWAYS">
-            <columns>
-               <TableColumn fx:id="columnTitle" editable="false" minWidth="-Infinity" prefWidth="-1.0" resizable="false" sortable="false" text="Titre" />
-               <TableColumn fx:id="columnAuthor" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false" sortable="false" text="Artiste" />
-               <TableColumn fx:id="columnAlbum" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false" sortable="false" text="Album" />
-               <TableColumn fx:id="columnActions" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false" sortable="false" text="Actions" />
-            </columns>
-            <VBox.margin>
-               <Insets top="25.0" />
-            </VBox.margin>
-            <columnResizePolicy>
-               <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-            </columnResizePolicy>
-         </TableView>
-      </VBox>
    </center>
    <right>
       <VBox alignment="TOP_CENTER" minWidth="250.0" styleClass="h1" BorderPane.alignment="CENTER">
