+++ src/main/resources/fxml/my-track-view.fxml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox xmlns:fx="http://javafx.com/fxml"
+      xmlns="http://javafx.com/javafx" alignment="TOP_CENTER"
+      style="-fx-border-width: 1; -fx-border-color: gray; -fx-border-style: hidden solid hidden solid;"
+      BorderPane.alignment="CENTER">
+    <padding>
+        <Insets left="25.0" right="25.0"/>
+    </padding>
+    <Label styleClass="h1" text="Vos musiques"/>
+    <Separator prefWidth="200.0" styleClass="separator-dark"/>
+    <TableView fx:id="trackList" VBox.vgrow="ALWAYS">
+        <columns>
+            <TableColumn fx:id="columnTitle" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false"
+                         text="Titre"/>
+            <TableColumn fx:id="columnAuthor" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false"
+                         text="Artiste"/>
+            <TableColumn fx:id="columnAlbum" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false"
+                         text="Album"/>
+            <TableColumn fx:id="columnActions" editable="false" minWidth="-1.0" prefWidth="-1.0" resizable="false"
+                         text="Actions"/>
+        </columns>
+        <VBox.margin>
+            <Insets top="25.0"/>
+        </VBox.margin>
+        <columnResizePolicy>
+            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY"/>
+        </columnResizePolicy>
+    </TableView>
+</VBox>
